Moved navigator annotation => org.vaadin.spring.navigator.annotation
