+++ samples/mvp-sample/src/main/java/org/vaadin/spring/samples/mvp/ui/component/listener/TabSelectedListener.java
@@ -9,7 +9,7 @@ import org.springframework.context.ApplicationContext;
 import org.vaadin.spring.annotation.VaadinComponent;
 import org.vaadin.spring.annotation.VaadinUIScope;
 import org.vaadin.spring.events.EventBus;
-import org.vaadin.spring.navigator.VaadinPresenter;
+import org.vaadin.spring.navigator.annotation.VaadinPresenter;
 import org.vaadin.spring.navigator.annotation.VaadinView;
 import org.vaadin.spring.samples.mvp.ui.component.util.ControlsContext;
 import org.vaadin.spring.samples.mvp.ui.presenter.Screen;
