+++ samples/mvp-sample/src/main/java/org/vaadin/spring/samples/mvp/ui/component/selector/AnyParticipantSelectorPresenter.java
@@ -4,7 +4,7 @@ import org.joda.time.DateTime;
 import org.vaadin.spring.events.Event;
 import org.vaadin.spring.events.annotation.EventBusListenerMethod;
 import org.vaadin.spring.navigator.Presenter;
-import org.vaadin.spring.navigator.VaadinPresenter;
+import org.vaadin.spring.navigator.annotation.VaadinPresenter;
 import org.vaadin.spring.samples.mvp.ui.component.listener.MarketDaySelectedListener;
 import org.vaadin.spring.samples.mvp.util.SSTimeUtil;
