+++ samples/mvp-sample/src/main/java/org/vaadin/spring/samples/mvp/ui/presenter/BannerPresenter.java
@@ -6,7 +6,7 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.userdetails.User;
 import org.vaadin.spring.events.annotation.EventBusListenerMethod;
 import org.vaadin.spring.navigator.Presenter;
-import org.vaadin.spring.navigator.VaadinPresenter;
+import org.vaadin.spring.navigator.annotation.VaadinPresenter;
 import org.vaadin.spring.samples.mvp.ui.view.BannerView;
 import org.vaadin.spring.security.VaadinSecurity;
