+++ samples/mvp-sample/src/main/java/org/vaadin/spring/samples/mvp/ui/presenter/BodyPresenter.java
@@ -6,7 +6,7 @@ import org.vaadin.spring.events.Event;
 import org.vaadin.spring.events.EventScope;
 import org.vaadin.spring.events.annotation.EventBusListenerMethod;
 import org.vaadin.spring.navigator.Presenter;
-import org.vaadin.spring.navigator.VaadinPresenter;
+import org.vaadin.spring.navigator.annotation.VaadinPresenter;
 import org.vaadin.spring.samples.mvp.ui.view.BodyView;
 
 @VaadinPresenter(viewName = BodyView.NAME)
