+++ samples/mvp-sample/src/main/java/org/vaadin/spring/samples/mvp/ui/presenter/NavigationPanelPresenter.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.vaadin.spring.events.Event;
 import org.vaadin.spring.events.annotation.EventBusListenerMethod;
 import org.vaadin.spring.navigator.Presenter;
-import org.vaadin.spring.navigator.VaadinPresenter;
+import org.vaadin.spring.navigator.annotation.VaadinPresenter;
 import org.vaadin.spring.samples.mvp.ui.component.nav.NavElement;
 import org.vaadin.spring.samples.mvp.ui.component.nav.NavElementFactory;
 import org.vaadin.spring.samples.mvp.ui.view.NavigationPanelView;
