+++ samples/mvp-sample/src/main/java/org/vaadin/spring/samples/mvp/ui/presenter/TabPanelPresenter.java
@@ -2,7 +2,7 @@ package org.vaadin.spring.samples.mvp.ui.presenter;
 
 import org.vaadin.spring.events.annotation.EventBusListenerMethod;
 import org.vaadin.spring.navigator.Presenter;
-import org.vaadin.spring.navigator.VaadinPresenter;
+import org.vaadin.spring.navigator.annotation.VaadinPresenter;
 import org.vaadin.spring.samples.mvp.ui.component.nav.NavElement;
 import org.vaadin.spring.samples.mvp.ui.view.TabPanelView;
