+++ spring-vaadin-mvp/src/test/java/org/vaadin/spring/navigator/FooPresenter.java
@@ -2,6 +2,7 @@ package org.vaadin.spring.navigator;
 
 import org.vaadin.spring.annotation.VaadinUIScope;
 import org.vaadin.spring.events.annotation.EventBusListenerMethod;
+import org.vaadin.spring.navigator.annotation.VaadinPresenter;
 
 @VaadinUIScope
 @VaadinPresenter(viewName=FooView.NAME)
